requirements.txt
apt-xapian-index==0.47
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
Brlapi==0.6.6
builtwith==1.3.3
certifi==2018.1.18
chardet==3.0.4
click==7.1.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.4
djangorestframework==3.11.0
entrypoints==0.3
Flask==1.1.1
httplib2==0.9.2
idna==2.6
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
notebook==6.0.3
numpy==1.18.2
oauth==1.0.1
olefile==0.45.1
pandocfilters==1.4.2
parso==0.6.2
pbkdf2==1.3
persepolis==3.2.0
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.0.0
psutil==5.4.2
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyftpdlib==1.5.6
Pygments==2.5.2
pygobject==3.26.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyrsistent==0.15.7
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2019.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.14.0
sqlparse==0.3.1
ssh-import-id==5.7
sympy==1.5.1
system-service==0.3
systemd-python==234
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
vboxapi==1.0
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wifi==0.3.8
wrapt==1.11.2
xkit==0.0.0
youtube-dl==2018.3.14
zipp==3.1.0
zope.interface==4.3.2
